test(park-keeper): MiniMap 千行測試檔按功能拆分與羅盤導航 E2E 旅程
- MiniMap.test.tsx（1022 行）拆為 config/behavior/render/photo 四檔
- 共用 mock 抽至 miniMapTestKit（motion/react-leaflet/leaflet/mockTheme 單一來源）
- 補 photoOffset 初始套用與預設偏移單元測試（持久化重開套用驗證）
- marker 徽章配色參數改必填，移除硬編預設值
- 新增 navigation-compass E2E：記錄→開導航→羅盤渲染→heading 注入→關閉

測試：vitest 元件測試 11 檔 116 測試全綠；Playwright navigation-compass mobile 通過
